Read file inputs are now properly stored somewhere; basic data types added; expected inputs more specific
File inputs are now stored in a Pack data type. Basic operations for this data type have been added.
Right now it uses a simple list to keep track of its rectangles, though this should be changed in the future.
Possibly transform it into an ADT with an overridable addRectangle(..) and getNumberOfRectangles(..) methods.

Rectangle data type added. Stores an ID, width, and height. Could (or rather, should) be extended with rotations and position on the grid.

Expected Strings from the input are now more specific. An error will be thrown if the parsed data does not match the expected data. This is not yet complete with strings that are important for the rest of the input.

See the TODOs for more information.