GiantJavaFile completed
The giantJavaFile was completed and handed in to Momoter (with success).
As mentioned in an Announcement on Canvas, test case 8 and 9 are currently broken (and fail).

Momoter expects a file named PackingSolver.java. All other classes are pasted below this file's main class, but without their public keywords.
the PackingSolver class is the only class with a main method.