Testcases for InputReader finished + OutputWriter changes + bugfixes
The testcases for the InputReader are up and running. All tests were passed.

OutputWriter now takes a PrintStream as a parameter so that any PrintStream can be chosen as opposed to just System.out.

Fixed a bug where the number of rectangles of packdata depended on the containerHeight.
Fixed a bug where the containerHeight of packData could be less than -1