Basic GUI + ValidCheck update + bugfixes
A basic GUI was added. It shows the rectangles from a Pack on the screen.
It's currently a WIP, but the base functionality is there.

ValidCheck's methods have been made static, have had a name update, and have had their parameters swapped to more closely resemble their new names.

Bugfixes and minor changes:
- Fixed a bug where input.txt generated by InputGenerator was being tracked by git.
- Fixed a bug where MomotorMerger would try to merge GUI and InputGenerator.
- Added missing class description to GUI and InputGenerator.
- Renamed nl.tue.algorithms.dbl.wholeFile to nl.tue.algorithms.dbl.utilities
- Fixed a bug with unrecognised imports in FirstFitDecreasingHeightTest