GUI update + container size methods + test files moved
Container size is now shown in the GUI (in the title next to the coverage area).

The Background colour of the GUI was changed to a very light yellow colour. Rectangles can now be Orange as well. Rectangles can no longer be white or light grey.

Added a method in Algorithm that calculates its pack's container size. This can (and should) be overwritten by its subclasses if deemed necessary for runtime. This has already been done for the FFDW algorithm.

Test file moved out of the test-package into the /res/globalTests folder.
.gitignore updated as to ignore /res/localTests so that unimportant updates to one of your local test files is not tracked by git.