Bugfixes
Fixed an issue where, due to integer overflow, an incorrect container area was calculated. In cases of Integer Overflow, INTEGER.MAX_VALUE is now taken as Container Area instead.

Fixed an issue where a fixed height of 0 would indicate that there was a fixed height.

Fixed an issue in FFDW when it tried to run on large inputs without fixed height. It now instead does not run when there is no fixed height.

Improved memory performance of CompoundAlgorithm, where Packs are now only created before an Algorithm is invoked. The Java Garbage Collector can then take care of the Pack of the previously invoked Algorithm.

It seems that Java runs out of Heap Space Memory when trying to run the GUI with the FirstFitDecreasingHeight Algorithm and large inputs. There is no fix for this yet.