Rotations based on a Rectangle's Ratio
Rectangles in a PackWidthQueue and PackLinkedList are now rotated if their width-height ratio exceeds 3.0 (that is, the height is at least 3 times as big as the width)

Added the BruteForce testcase that seems to have odd interactions with the pruning.
For some reason, rotations do not work in BruteForce either.

Rotated Rectangles now have an 'R' in their bottomleft corner in the GUI.