CompoundAlgorithm Bugfix
Fixed an issue where CompoundAlgorithm created PackData that had rotations enabled when it shouldn't (and vice versa).

Re-enabled debug prints.

Added a warning when rotations are disallowed but a rectangle exceeds the fixed height. (The rectangle will still be rotated though).

Added 2 testcases that were found to be buggy when using CompoundAlgirhtm (invoking BinaryPacker)