Rotate Ratio Bugfixes
Fixed an issue where rectangles would be rotated (due to a valid ratio) while their Width exceeded the fixed height of a Pack.

A warning is now shown when a rectangle's height AND width both exceed the fixed height (in cases where rotations are allowed). Rectangles that have this problem are still rotated.

Fixed an issue where a Pack's height was not calculated correctly.

Fixed an issue in CompoundAlgorithm where it would copy over the entire pack from the best invoked algorithm, rather than the rectangle placements only.
As such, the GUI wil no longer show an error if the best algorithm has no rotations apart from those that exceed the fixed height.