Comparator Bugfix + Exception hiding
Fixed an issue in all Comparators where in rare cases it threw an Exception if two rectangles had the exact same size.

When an invoked algorithm by CompoundAlgorithm throws an exception, it is now ignored when outside debug mode.
The invoked algorithm's "solution" will have Integer.MAX_VALUE as an area.