averageCalculator up and running
The averageCalculator is up and running. It executes all testcases in a specific folder (res/averageCalculator/input by default), and prints the results as a .csv-file (comma separated values).
This .csv file can be easily opened in programs such as Microsoft Excel (for instance)
Also outputs an R-file with the outputs as a dataset.
Either R or Microsoft Excel can then be used to calculate things such as mean, or to make fancy graphs.